pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "Livrando"
version = "1.0.0"
description = "Organizador automático de livros digitais"
readme = "README.md"
requires-python = ">=3.8"
license = {file = "LICENSE"}
authors = [
    {name = "vjuh", email = "vjuh@hotmail.com"}
]
keywords = ["ebook", "organizer", "metadata", "books", "epub", "pdf"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: End Users/Desktop",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Utilities",
]

dependencies = [
    "requests>=2.28.0",
    "ebooklib>=0.17.1",
    "PyPDF2>=2.10.0",
    "Pillow>=9.0.0"
]

[project.scripts]
Livrando = "Livrando:main"

[project.urls]
Homepage = "https://github.com/vjuh/Livrando"
Issues = "https://github.com/vjuh/Livrando/issues"
Source = "https://github.com/vjuh/Livrando"

[tool.setuptools]
packages = ["Livrando"]

[tool.setuptools.package-dir]
Livrando = "."